app/requirements.txt
dateparser
feedparser
fuzzywuzzy
html2markdown
python-Levenshtein
python-telegram-bot
